@@ -15,7 +15,7 @@ namespace NoSleepApp {
             Application.SetHighDpiMode(HighDpiMode.SystemAware);
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new SleepForm());
         }
     }
 }
